requirements.txt
numpy
torch
transformers
sentence_transformers
gradio
pymupdf
openai
faiss-cpu
requests
beautifulsoup4